requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.74.1
anyio==4.11.0
asyncio-throttle==1.0.2
attrs==25.4.0
Authlib==1.6.5
backoff==2.2.1
base58==2.1.1
beartype==0.22.6
beautifulsoup4==4.14.2
bitarray==2.9.2
blinker==1.9.0
boto3==1.35.99
botocore==1.35.99
build==1.3.0
cachetools==6.2.2
ccxt==4.5.20
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
coincurve==21.0.0
cryptography==46.0.3
cyclopts==4.2.5
cytoolz==1.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.3
email-validator==2.3.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
Events==0.5
exceptiongroup==1.3.1
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
fastmcp==2.13.1
Flask==3.1.2
frozenlist==1.8.0
google==3.0.0
google-api-core==2.28.1
google-auth==2.43.0
google-genai==1.52.0
googleapis-common-protos==1.72.0
gql==4.0.0
graphql-core==3.2.7
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.22.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.0
neo-mamba==3.1.0
neo3crypto==0.4.4
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==2.3.5
openai==2.8.1
openapi-pydantic==0.5.1
orjson==3.11.4
packaging==24.2
pandas==2.3.3
parsimonious==0.10.0
pathable==0.4.4
pathvalidate==3.3.1
pinecone==8.0.0
pinecone-plugin-assistant==3.0.1
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybiginteger==1.3.4
pybiginteger-stubs==1.3.4
pycares==4.11.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyperclip==1.11.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyunormalize==17.0.0
PyYAML==6.0.3
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-rst==1.3.2
rich-toolkit==0.16.0
rignore==0.7.6
rlp==4.1.0
rpds-py==0.29.0
rsa==4.9.1
s3transfer==0.10.4
sentry-sdk==2.45.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
spoon-ai-sdk==0.3.3
spoon-toolkits==0.2.2
SQLAlchemy==2.0.44
sse-starlette==3.0.3
starlette==0.50.0
TA-Lib==0.6.8
tenacity==9.1.2
termcolor==3.2.0
tiktoken==0.12.0
toolz==1.1.0
tqdm==4.67.1
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
web3==7.11.0
websockets==15.0.1
Werkzeug==3.1.3
x402==0.2.1
yarl==1.22.0
